pyproject.toml
# Copyright 2026 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "backend"
version = "0.1.0"
description = "Backend services and agent simulations for Race Condition"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.26,<0.4.0",
    "asyncpg>=0.30.0",
    "google-adk[extensions,eval]==1.30.0",
    "google-cloud-iamconnectorcredentials>=0.1.0",
    "google-cloud-aiplatform>=1.143.0",
    "google-adk-community>=0.4.1",
    "google-genai>=1.70.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "google-cloud-pubsub>=2.30.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "protobuf>=5.29.3,<7",
    "grpcio-tools>=1.70.0",
    "websockets",
    "redis>=5.0.0",
    "anyio>=4.8.0",
    "pypdf>=6.10.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0",
    "diff-cover>=9.0",
    "ruff>=0.8.0",
    "honcho>=1.1.0",
    "respx>=0.22.0",
]

[project.scripts]
init = "scripts.core.sim:init"
start = "scripts.core.sim:start_cli"
stop = "scripts.core.sim:stop_cli"
restart = "scripts.core.sim:restart_cli"
test = "scripts.core.sim:test"
test-integration = "scripts.core.sim:test_integration"
deploy = "scripts.deploy.deploy:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
managed = true
package = true

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.pytest.ini_options]
addopts = "-m 'not slow and not integration'"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["agents", "scripts/deploy"]
python_files = ["*_test.py", "test_*.py"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services (Redis, PubSub)",
]

[tool.coverage.run]
source = ["agents"]
omit = ["agents/tests/*", "agents/**/test_*"]

[tool.coverage.report]
fail_under = 60
show_missing = true

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W"]

[tool.ruff.lint.per-file-ignores]
# ADK agents use late imports for App/card registration (intentional pattern)
"agents/*/agent.py" = ["E402"]
"agents/*/*/agent.py" = ["E402"]
# __init__.py files re-export symbols
"agents/**/__init__.py" = ["F401"]
# Utility with circular dep avoidance and tests with dynamic module loading
"agents/utils/a2a.py" = ["E402"]
"agents/tests/*.py" = ["E402"]

